djangoreactapi/backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-queryset-csv==1.1.0
djangorestframework==3.12.2
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
mysqlclient==2.0.3
numpy==1.19.5
opencv-python==4.5.1.48
pafy==0.5.5
pandas==1.1.5
Pillow==8.2.0
py-bcrypt==0.4
PyJWT==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.60.0
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.4
wincertstore==0.2
youtube-dl==2021.4.17
